Move data from audits to lib to fix runner-test failure
