+++ lighthouse-core/lib/large-javascript-libraries/bundlephobia-database.json
