+++ lighthouse-core/lib/large-javascript-libraries/library-suggestions.js
